DL-14089 Test suite missed
