+++ conf/application.conf
@@ -20,7 +20,7 @@ play.modules.enabled += "config.Wiring"
 
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.HttpClientV2Module"
 
-play.http.errorHandler = "config.ErrorHandler"
+play.http.errorHandler = "config.BCHandler"
 play.filters.enabled += play.filters.csp.CSPFilter
 
 # The application languages
